feat(logo): wip
